requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
duckduckgo-search==3.9.6
openai==1.12.0
thefuzz==0.22.1
python-Levenshtein==0.23.0
colorama==0.4.6
json5==0.9.14
Flask==3.0.3
openpyxl==3.1.5
httpx==0.27.2
